fix(ci): unblock Pest marking checkout and normalize stocks decimals
Harden MarkingAndEgais happy-path CIS (HTML-safe serial), force marking mock via config in tests/CI, and align stocks.actual/reserved/available to decimal(14,3) end-to-end so remote main can go green after push.

+++ database/migrations/2026_08_01_000011_create_stocks_table.php
@@ -21,9 +21,9 @@ return new class extends Migration
             $table->foreignId('tenant_id')->constrained()->cascadeOnDelete();
             $table->foreignId('warehouse_id')->constrained()->cascadeOnDelete();
             $table->foreignId('product_id')->constrained('products_services')->cascadeOnDelete();
-            $table->unsignedInteger('actual')->default(0);
-            $table->unsignedInteger('reserved')->default(0);
-            $table->unsignedInteger('available')->default(0);
+            $table->decimal('actual', 14, 3)->default(0);
+            $table->decimal('reserved', 14, 3)->default(0);
+            $table->decimal('available', 14, 3)->default(0);
             $table->timestampsTz();
             $table->unique(['tenant_id', 'warehouse_id', 'product_id']);
             $table->index('tenant_id');
